Make flake.nix a bit cleaner
Moved paths to modules to the top of module lists.
Switched to using `builtins.head` for `nixosModules.home-manager` instead of mapping.
Switched to using `lib.genAttrs` for `homeConfigurations.${user}.${host}` instead of mapping a list of attrs.
Switched to using the nixos definition for home-manager in `homeConfigurations` (they were basically just repeated).